Gameで管理していたstateをAppで管理するように
