+++ src/components/Game.js
@@ -1,23 +1,9 @@
-import { useState } from 'react';
 import '../App.css';
 import Board from './Board';
+import { useState } from 'react';
 
-function Game() {
-  const [history, setHistory] = useState([Array(9).fill(null)]);
-  const [currentMove, setCurrentMove] = useState(0);
-  const xIsNext = currentMove % 2 === 0;
-  const currentSquares = history[currentMove];
-
-  function handlePlay(nextSquares) {
-    const nextHistory = [...history.slice(0, currentMove + 1), nextSquares];
-    setHistory(nextHistory);
-    setCurrentMove(nextHistory.length - 1);
-  }
-
-  function jumpTo(nextMove) {
-    setCurrentMove(nextMove);
-  }
-
+function Game({history, xIsNext, currentSquares, handlePlay, jumpTo}) {
+  
   const moves = history.map((squares, move) => {
     let description;
     if (move > 0) {
